Add new migrations for len_surveyed

+++ src/api/models/view_models/benthicpit.py
@@ -74,9 +74,7 @@ CREATE OR REPLACE VIEW public.vw_benthicpit_obs
     observers = JSONField(null=True, blank=True)
     transect_number = models.PositiveSmallIntegerField()
     label = models.CharField(max_length=50, blank=True)
-    transect_len_surveyed = models.PositiveSmallIntegerField(
-        verbose_name=_("transect length surveyed (m)")
-    )
+    transect_len_surveyed = models.DecimalField(max_digits=3, decimal_places=1, verbose_name=_("transect length surveyed (m)"))
     depth = models.DecimalField(
         max_digits=3, decimal_places=1, verbose_name=_("depth (m)")
     )
@@ -160,9 +158,7 @@ INNER JOIN (
 
     observers = JSONField(null=True, blank=True)
     transect_number = models.PositiveSmallIntegerField()
-    transect_len_surveyed = models.PositiveSmallIntegerField(
-        verbose_name=_("transect length surveyed (m)")
-    )
+    transect_len_surveyed = models.DecimalField(max_digits=3, decimal_places=1, verbose_name=_("transect length surveyed (m)"))
     depth = models.DecimalField(
         max_digits=3, decimal_places=1, verbose_name=_("depth (m)")
     )
